requirements.txt
sqlalchemy>=2.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
cryptography>=41.0.0
apscheduler>=3.10.0
streamlit>=1.26.0
plotly>=5.18.0
matplotlib>=3.7.0
